requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cssselect2==0.4.1
decorator==5.0.9
defusedxml==0.6.0
Django==3.1.5
django-appconf==1.0.4
django-ckeditor==6.0.0
django-cropper-image==1.0.5
django-easy-pdf==0.1.1
django-file-download==1.0.2
django-image-cropping==1.6.1
django-imagekit==4.0.2
django-imagekit-cropper==1.17
django-js-asset==1.2.2
django-json-response==1.1.5
django-logit==1.1.3
django-phonenumber-field==5.1.0
django-response==1.1.1
django-storages==1.11.1
django-taggit==1.3.0
easy-thumbnails==2.7.1
future==0.18.2
html5lib==1.1
idna==2.10
phonenumbers==8.12.22
pilkit==2.0
Pillow==8.1.0
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.5
q==2.6
reportlab==3.5.59
requests==2.25.1
self==2020.12.3
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
TimeConvert==1.5.2
tinycss2==1.1.0
tzlocal==2.1
urllib3==1.26.3
WeasyPrint==52.2
webencodings==0.5.1
xhtml2pdf==0.2.5
gunicorn
psycopg2-binary
django-environ
